implemented listen_to_seller with db and optimized random things

+++ sql/market.sql
@@ -3,7 +3,7 @@ PRAGMA foreign_keys = ON;
 CREATE TABLE IF NOT EXISTS product (
     product_name TEXT,
     seller_id INTEGER,
-    price NUMERIC(10, 2),
+    price INTEGER,
     quantity INTEGER,
     PRIMARY KEY (product_name, seller_id)
 );
